reassign duties of creating booth list from main BoothMap to Booth component and changed name to Booths
