+++ client/src/BoothMap/components/DeleteButton.js
@@ -1,7 +1,7 @@
 import React, {memo} from 'react'
 import deleteBooth from '../functions/databaseOperations/deleteBooth'
 import trashIcon from '../../assets/Icons/trash.png'
-import { StyledDeleteButton } from './styledComponents'
+import { StyledDeleteButton } from '../styledComponents'
 
 export default memo(function DeleteButton({ selected, setBooths }) {
     const handleClick = () => {
