+++ client/src/BoothMap/functions/databaseOperations/getBooths.js
@@ -1,9 +1,9 @@
 import boothsRef from './firestore'
 import handleErrors from '../errorHandler'
 
-const dummyData = [{id:'A01',top:300,left:300},{id:'A02',top:200,left:300}]
 
 export default function getBooths(setBooths) {
+    // const dummyData = [{id:'A01',top:300,left:300},{id:'A02',top:200,left:300},{id:'A03',top:100,left:300}]
     // setBooths(dummyData)
     boothsRef.get()
         .then(res => {
