+++ client/src/BoothMap/index.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react'
-import { Container } from './components/styledComponents/'
-import Booth from './components/Booth/'
+import { Container } from './styledComponents'
+import Booths from './components/Booths'
 import AddBoothForm from './components/AddBoothForm'
 import DeleteButton from './components/DeleteButton'
 import reserveBooth from './functions/databaseOperations/reserveBooth'
@@ -20,21 +20,15 @@ export default function BoothMap({ onChange, ADMIN }) {
 
     const boothIsSelected = booths.some(booth => booth.id === state.selected)
 
-    const Booths = booths.map(doc => <Booth
-        key={doc.id}
-        {...doc}
-        ADMIN={ADMIN}
-        selected={state.selected === doc.id} />)
-
     const changeSelectedBooth = e => {
         e.target.textContent !== state.selected &&
-            setState({ selected: e.target.textContent, isDragging: true })
+            setState({ selected: e.target.textContent })
     }
 
     return <Container ADMIN={ADMIN} onMouseDown={changeSelectedBooth}>
-        {Booths}
+        <Booths {...state} booths={booths}/>
         {ADMIN && <>
-            <AddBoothForm {...{ setBooths }} />
+            <AddBoothForm {...{ setBooths }}/>
             {boothIsSelected && <DeleteButton selected={state.selected} {...{ setBooths }} />}
         </>}
         {!ADMIN &&
