+++ client/src/BoothMap/styledComponents/Booth.js
