+++ client/src/BoothMap/styledComponents/index.js
